[ Widget Preview ] Refactor the flutter widget-preview command
No functional changes.
