+++ packages/flutter_tools/lib/src/commands/widget_preview.dart
@@ -5,13 +5,11 @@
 import 'dart:async';
 
 import 'package:args/args.dart';
-import 'package:meta/meta.dart';
 import 'package:package_config/package_config.dart';
 import 'package:process/process.dart';
 
 import '../artifacts.dart';
 import '../base/common.dart';
-import '../base/deferred_component.dart';
 import '../base/file_system.dart';
 import '../base/logger.dart';
 import '../base/os.dart';
@@ -20,20 +18,19 @@ import '../base/process.dart';
 import '../build_info.dart';
 import '../bundle.dart' as bundle;
 import '../cache.dart';
-import '../convert.dart';
-import '../dart/pub.dart';
 import '../device.dart';
-import '../flutter_manifest.dart';
 import '../globals.dart' as globals;
 import '../isolated/resident_web_runner.dart';
 import '../project.dart';
 import '../resident_runner.dart';
 import '../runner/flutter_command.dart';
 import '../runner/flutter_command_runner.dart';
+import '../widget_preview/dependency_graph.dart';
 import '../widget_preview/dtd_services.dart';
 import '../widget_preview/preview_code_generator.dart';
 import '../widget_preview/preview_detector.dart';
 import '../widget_preview/preview_manifest.dart';
+import '../widget_preview/preview_pubspec_builder.dart';
 import 'create_base.dart';
 
 class WidgetPreviewCommand extends FlutterCommand {
@@ -193,12 +190,20 @@ final class WidgetPreviewStartCommand extends WidgetPreviewSubCommandBase with C
 
   late final FlutterProject rootProject = getRootProject();
 
+  late final PreviewPubspecBuilder _previewPubspecBuilder = PreviewPubspecBuilder(
+    logger: logger,
+    verbose: verbose,
+    offline: offline,
+    rootProject: rootProject,
+    previewManifest: _previewManifest,
+  );
+
   late final PreviewDetector _previewDetector = PreviewDetector(
     projectRoot: rootProject.directory,
     logger: logger,
     fs: fs,
     onChangeDetected: onChangeDetected,
-    onPubspecChangeDetected: onPubspecChangeDetected,
+    onPubspecChangeDetected: _previewPubspecBuilder.onPubspecChangeDetected,
   );
 
   late final PreviewCodeGenerator _previewCodeGenerator;
@@ -282,7 +287,7 @@ final class WidgetPreviewStartCommand extends WidgetPreviewSubCommandBase with C
       //
       // Populate the pubspec after the initial build to avoid blowing away the package_config.json
       // which may have manual changes for flutter_gen support.
-      await _populatePreviewPubspec(rootProject: rootProject);
+      await _previewPubspecBuilder.populatePreviewPubspec(rootProject: rootProject);
     }
 
     final PreviewDependencyGraph graph = await _previewDetector.initialize();
@@ -310,12 +315,6 @@ final class WidgetPreviewStartCommand extends WidgetPreviewSubCommandBase with C
     _previewCodeGenerator.populatePreviewsInGeneratedPreviewScaffold(previews);
     logger.printStatus('Triggering reload based on change to preview set: $previews');
     _widgetPreviewApp?.restart();
-  }
-
-  void onPubspecChangeDetected() {
-    // TODO(bkonyi): trigger hot reload or restart?
-    logger.printStatus('Changes to pubspec.yaml detected.');
-    _populatePreviewPubspec(rootProject: rootProject);
   }
 
   /// Configures the Dart Tooling Daemon connection.
@@ -417,194 +416,6 @@ final class WidgetPreviewStartCommand extends WidgetPreviewSubCommandBase with C
     // WARNING: this log message is used by test/integration.shard/widget_preview_test.dart
     logger.printStatus('Done loading previews.');
     return runner;
-  }
-
-  @visibleForTesting
-  static const Map<String, String> flutterGenPackageConfigEntry = <String, String>{
-    'name': 'flutter_gen',
-    'rootUri': '../../flutter_gen',
-    'languageVersion': '2.12',
-  };
-
-  /// Maps asset URIs to relative paths for the widget preview project to
-  /// include.
-  @visibleForTesting
-  static Uri transformAssetUri(Uri uri) {
-    // Assets provided by packages always start with 'packages' and do not
-    // require their URIs to be updated.
-    if (uri.path.startsWith('packages')) {
-      return uri;
-    }
-    // Otherwise, the asset is contained within the root project and needs
-    // to be referenced from the widget preview scaffold project's pubspec.
-    return Uri(path: '../../${uri.path}');
-  }
-
-  @visibleForTesting
-  static AssetsEntry transformAssetsEntry(AssetsEntry asset) {
-    return AssetsEntry(
-      uri: transformAssetUri(asset.uri),
-      flavors: asset.flavors,
-      transformers: asset.transformers,
-    );
-  }
-
-  @visibleForTesting
-  static FontAsset transformFontAsset(FontAsset asset) {
-    return FontAsset(transformAssetUri(asset.assetUri), weight: asset.weight, style: asset.style);
-  }
-
-  @visibleForTesting
-  static DeferredComponent transformDeferredComponent(DeferredComponent component) {
-    return DeferredComponent(
-      name: component.name,
-      // TODO(bkonyi): verify these library paths are always package: paths from the parent project.
-      libraries: component.libraries,
-      assets: component.assets.map(transformAssetsEntry).toList(),
-    );
-  }
-
-  @visibleForTesting
-  FlutterManifest buildPubspec({
-    required FlutterManifest rootManifest,
-    required FlutterManifest widgetPreviewManifest,
-  }) {
-    final List<AssetsEntry> assets = rootManifest.assets.map(transformAssetsEntry).toList();
-
-    final List<Font> fonts = <Font>[
-      ...widgetPreviewManifest.fonts,
-      ...rootManifest.fonts.map((Font font) {
-        return Font(font.familyName, font.fontAssets.map(transformFontAsset).toList());
-      }),
-    ];
-
-    final List<Uri> shaders = rootManifest.shaders.map(transformAssetUri).toList();
-
-    final List<DeferredComponent>? deferredComponents =
-        rootManifest.deferredComponents?.map(transformDeferredComponent).toList();
-
-    return widgetPreviewManifest.copyWith(
-      logger: logger,
-      assets: assets,
-      fonts: fonts,
-      shaders: shaders,
-      deferredComponents: deferredComponents,
-    );
-  }
-
-  Future<void> _populatePreviewPubspec({required FlutterProject rootProject}) async {
-    final FlutterProject widgetPreviewScaffoldProject = rootProject.widgetPreviewScaffoldProject;
-
-    // Overwrite the pubspec for the preview scaffold project to include assets
-    // from the root project.
-    widgetPreviewScaffoldProject.replacePubspec(
-      buildPubspec(
-        rootManifest: rootProject.manifest,
-        widgetPreviewManifest: widgetPreviewScaffoldProject.manifest,
-      ),
-    );
-
-    // Adds a path dependency on the parent project so previews can be
-    // imported directly into the preview scaffold.
-    const String pubAdd = 'add';
-    // Use `json.encode` to handle escapes correctly.
-    final String pathDescriptor = json.encode(<String, Object?>{
-      // `pub add` interprets relative paths relative to the current directory.
-      'path': rootProject.directory.fileSystem.path.relative(rootProject.directory.path),
-    });
-
-    final PubOutputMode outputMode = verbose ? PubOutputMode.all : PubOutputMode.failuresOnly;
-    await pub.interactively(
-      <String>[
-        pubAdd,
-        if (offline) '--offline',
-        '--directory',
-        widgetPreviewScaffoldProject.directory.path,
-        // Ensure the path using POSIX separators, otherwise the "path_not_posix" check will fail.
-        '${rootProject.manifest.appName}:$pathDescriptor',
-      ],
-      context: PubContext.pubAdd,
-      command: pubAdd,
-      touchesPackageConfig: true,
-      outputMode: outputMode,
-    );
-
-    // Adds dependencies on:
-    //   - dtd, which is used to connect to the Dart Tooling Daemon to establish communication
-    //     with other developer tools.
-    //   - flutter_lints, which is referenced by the analysis_options.yaml generated by the 'app'
-    //     template.
-    //   - stack_trace, which is used to generate terse stack traces for displaying errors thrown
-    //     by widgets being previewed.
-    //   - url_launcher, which is used to open a browser to the preview documentation.
-    await pub.interactively(
-      <String>[
-        pubAdd,
-        if (offline) '--offline',
-        '--directory',
-        widgetPreviewScaffoldProject.directory.path,
-        'dtd',
-        'flutter_lints',
-        'stack_trace',
-        'url_launcher',
-      ],
-      context: PubContext.pubAdd,
-      command: pubAdd,
-      touchesPackageConfig: true,
-      outputMode: outputMode,
-    );
-
-    // Generate package_config.json.
-    await pub.get(
-      context: PubContext.create,
-      project: widgetPreviewScaffoldProject,
-      offline: offline,
-      outputMode: outputMode,
-    );
-
-    maybeAddFlutterGenToPackageConfig(rootProject: rootProject);
-    _previewManifest.updatePubspecHash();
-  }
-
-  /// Manually adds an entry for package:flutter_gen to the preview scaffold's
-  /// package_config.json if the target project makes use of localization.
-  ///
-  /// The Flutter Tool does this when running a Flutter project with
-  /// localization instead of modifying the user's pubspec.yaml to depend on it
-  /// as a path dependency. Unfortunately, the preview scaffold still needs to
-  /// add it directly to its package_config.json as the generated package name
-  /// isn't actually flutter_gen, which pub doesn't really like, and using the
-  /// actual package name will break applications which import
-  /// package:flutter_gen.
-  @visibleForTesting
-  void maybeAddFlutterGenToPackageConfig({required FlutterProject rootProject}) {
-    // TODO(matanlurey): Remove this once flutter_gen is removed.
-    //
-    // This is actually incorrect logic; the presence of a `generate: true`
-    // does *NOT* mean that we need to add `flutter_gen` to the package config,
-    // and never did, but the name of the manifest field was labeled and
-    // described incorrectly.
-    //
-    // Tracking removal: https://github.com/flutter/flutter/issues/102983.
-    if (!rootProject.manifest.generateLocalizations) {
-      return;
-    }
-    final FlutterProject widgetPreviewScaffoldProject = rootProject.widgetPreviewScaffoldProject;
-    final File packageConfig = widgetPreviewScaffoldProject.packageConfig;
-    final String previewPackageConfigPath = packageConfig.path;
-    if (!packageConfig.existsSync()) {
-      throw StateError(
-        "Could not find preview project's package_config.json at "
-        '$previewPackageConfigPath',
-      );
-    }
-    final Map<String, Object?> packageConfigJson =
-        json.decode(packageConfig.readAsStringSync()) as Map<String, Object?>;
-    (packageConfigJson['packages'] as List<dynamic>?)!.cast<Map<String, String>>().add(
-      flutterGenPackageConfigEntry,
-    );
-    packageConfig.writeAsStringSync(json.encode(packageConfigJson));
-    logger.printStatus('Added flutter_gen dependency to $previewPackageConfigPath');
   }
 }
